+++ cai_causal_graph/time_series_causal_graph.py
@@ -69,8 +69,8 @@ class TimeSeriesCausalGraph(CausalGraph):
     _NodeCls: Type[TimeSeriesNode] = TimeSeriesNode
     _EdgeCls: Type[Edge] = Edge
     _SummaryGraphCls: Type[CausalGraph] = CausalGraph
-    lag_to_nodes: Dict[int, List[TimeSeriesNode]]
-    variable_name_to_nodes: Dict[str, List[TimeSeriesNode]]
+    _lag_to_nodes: Dict[int, List[TimeSeriesNode]]
+    _variable_name_to_nodes: Dict[str, List[TimeSeriesNode]]
 
     def __init__(
         self,
@@ -115,8 +115,8 @@ class TimeSeriesCausalGraph(CausalGraph):
             not connected by edges.
         """
         # Initialize caches for fast lookups
-        self.lag_to_nodes = defaultdict(list)
-        self.variable_name_to_nodes = defaultdict(list)
+        self._lag_to_nodes = defaultdict(list)
+        self._variable_name_to_nodes = defaultdict(list)
         super().__init__(input_list, output_list, fully_connected)
 
         # list of variables in the graph, i.e. discarding the lags (X1(t-1) and X1 are the same variable)
@@ -728,7 +728,7 @@ class TimeSeriesCausalGraph(CausalGraph):
 
         identifier = self._check_node_exists(node)
         self._nodes_by_identifier[identifier] = node
-        self.add_node_to_cache(node)
+        self._add_node_to_cache(node)
 
         return node
 
@@ -778,7 +778,7 @@ class TimeSeriesCausalGraph(CausalGraph):
         """
         node = self.get_node(self._NodeCls.identifier_from(identifier))
         assert isinstance(node, TimeSeriesNode)  # for linting
-        self.remove_node_from_cache(node)
+        self._remove_node_from_cache(node)
         super().delete_node(identifier)
 
     @_reset_ts_graph_attributes
@@ -1289,29 +1289,32 @@ class TimeSeriesCausalGraph(CausalGraph):
         assert graph.variables is not None
         return adjacency_matrices, graph.variables
 
-    def add_node_to_cache(self, node: TimeSeriesNode):
-        """Add a node to the model caches."""
-        self.lag_to_nodes[node.time_lag].append(node)
-        self.variable_name_to_nodes[node.variable_name].append(node)
+    def _add_node_to_cache(self, node: TimeSeriesNode):
+        """Add a node to node caches."""
+        self._lag_to_nodes[node.time_lag].append(node)
+        self._variable_name_to_nodes[node.variable_name].append(node)
 
-    def remove_node_from_cache(self, node: TimeSeriesNode):
-        """Remove a node to the model caches."""
-        # Handle removal from lag-to-nodes cache
-        if node.time_lag in self.lag_to_nodes and node in self.lag_to_nodes[node.time_lag]:
-            self.lag_to_nodes[node.time_lag].remove(node)
-            # Optionally, clean up the entry if the list is now empty
-            if not self.lag_to_nodes[node.time_lag]:
-                del self.lag_to_nodes[node.time_lag]
+    def _remove_node_from_cache(self, node: TimeSeriesNode):
+        """Remove a node to node caches."""
+        try:
+            self._lag_to_nodes[node.time_lag].remove(node)
+            if not self._lag_to_nodes[node.time_lag]:
+                del self._lag_to_nodes[node.time_lag]
+        except ValueError as e:
+            raise ValueError(
+                f'Tried to remove node {node.identifier} from `TimeSeriesCausalGraph._lag_to_nodes` cache but the '
+                f'node was not found!'
+            ) from e
 
-        # Handle removal from variable-name-to-nodes cache
-        if (
-            node.variable_name in self.variable_name_to_nodes
-            and node in self.variable_name_to_nodes[node.variable_name]
-        ):
-            self.variable_name_to_nodes[node.variable_name].remove(node)
-            # Optionally, clean up the entry if the list is now empty
-            if not self.variable_name_to_nodes[node.variable_name]:
-                del self.variable_name_to_nodes[node.variable_name]
+        try:
+            self._variable_name_to_nodes[node.variable_name].remove(node)
+            if not self._variable_name_to_nodes[node.variable_name]:
+                del self._variable_name_to_nodes[node.variable_name]
+        except ValueError as e:
+            raise ValueError(
+                f'Tried to remove node {node.identifier} from `TimeSeriesCausalGraph.variable_name_to_nodes` cache '
+                f'but the node was not found!'
+            ) from e
 
     def get_nodes_at_lag(self, time_lag: int = 0) -> List[TimeSeriesNode]:
         """
@@ -1319,7 +1322,7 @@ class TimeSeriesCausalGraph(CausalGraph):
 
         :param time_lag: Time lag to return nodes for. Default is `0`.
         """
-        return self.lag_to_nodes[time_lag]
+        return self._lag_to_nodes[time_lag]
 
     def get_nodes_for_variable_name(self, variable_name: str) -> List[TimeSeriesNode]:
         """
@@ -1327,7 +1330,7 @@ class TimeSeriesCausalGraph(CausalGraph):
 
         :param variable_name: Variable name to return nodes for.
         """
-        return self.variable_name_to_nodes[variable_name]
+        return self._variable_name_to_nodes[variable_name]
 
     def get_contemporaneous_nodes(self, node: NodeLike) -> List[TimeSeriesNode]:
         """Return all nodes that are contemporaneous (i.e. have the same time_lag) to the provided node."""