+++ tests/test_time_series_graph.py
@@ -1871,7 +1871,7 @@ class TestTimeSeriesCausalGraphPrinting(unittest.TestCase):
         ts_graph = TimeSeriesCausalGraph(input_list=input_nodes, fully_connected=False)
 
         self.assertDictEqual(
-            dict(ts_graph.lag_to_nodes),
+            dict(ts_graph._lag_to_nodes),
             {0: [ts_graph['B']], -1: [ts_graph['A lag(n=1)']], -2: [ts_graph['A lag(n=2)']]},
         )
 
@@ -1882,11 +1882,11 @@ class TestTimeSeriesCausalGraphPrinting(unittest.TestCase):
         ts_graph.add_node('C')
 
         self.assertDictEqual(
-            dict(ts_graph.lag_to_nodes),
+            dict(ts_graph._lag_to_nodes),
             {0: [ts_graph['B'], ts_graph['C']], -1: [ts_graph['A lag(n=1)']], -2: [ts_graph['A lag(n=2)']]},
         )
         self.assertDictEqual(
-            dict(ts_graph.variable_name_to_nodes),
+            dict(ts_graph._variable_name_to_nodes),
             {'B': [ts_graph['B']], 'C': [ts_graph['C']], 'A': [ts_graph['A lag(n=2)'], ts_graph['A lag(n=1)']]},
         )
 
@@ -1896,9 +1896,9 @@ class TestTimeSeriesCausalGraphPrinting(unittest.TestCase):
         ts_graph = TimeSeriesCausalGraph(input_list=input_nodes, fully_connected=False)
 
         ts_graph.remove_node('A lag(n=2)')
-        self.assertDictEqual(dict(ts_graph.lag_to_nodes), {0: [ts_graph['B']], -1: [ts_graph['A lag(n=1)']]})
+        self.assertDictEqual(dict(ts_graph._lag_to_nodes), {0: [ts_graph['B']], -1: [ts_graph['A lag(n=1)']]})
         self.assertDictEqual(
-            dict(ts_graph.variable_name_to_nodes), {'B': [ts_graph['B']], 'A': [ts_graph['A lag(n=1)']]}
+            dict(ts_graph._variable_name_to_nodes), {'B': [ts_graph['B']], 'A': [ts_graph['A lag(n=1)']]}
         )
 
     def test_node_cache_update_deletion(self):
@@ -1907,9 +1907,9 @@ class TestTimeSeriesCausalGraphPrinting(unittest.TestCase):
         ts_graph = TimeSeriesCausalGraph(input_list=input_nodes, fully_connected=False)
 
         ts_graph.delete_node('A lag(n=2)')
-        self.assertDictEqual(dict(ts_graph.lag_to_nodes), {0: [ts_graph['B']], -1: [ts_graph['A lag(n=1)']]})
+        self.assertDictEqual(dict(ts_graph._lag_to_nodes), {0: [ts_graph['B']], -1: [ts_graph['A lag(n=1)']]})
         self.assertDictEqual(
-            dict(ts_graph.variable_name_to_nodes), {'A': [ts_graph['A lag(n=1)']], 'B': [ts_graph['B']]}
+            dict(ts_graph._variable_name_to_nodes), {'A': [ts_graph['A lag(n=1)']], 'B': [ts_graph['B']]}
         )
 
     def test_cache_when_from_causal_graph(self):
@@ -1919,7 +1919,15 @@ class TestTimeSeriesCausalGraphPrinting(unittest.TestCase):
         cg.add_edge('B', 'C')
         ts_graph = TimeSeriesCausalGraph.from_causal_graph(cg)
 
-        self.assertDictEqual(dict(ts_graph.lag_to_nodes), {0: [ts_graph['A'], ts_graph['B'], ts_graph['C']]})
+        self.assertDictEqual(dict(ts_graph._lag_to_nodes), {0: [ts_graph['A'], ts_graph['B'], ts_graph['C']]})
         self.assertDictEqual(
-            dict(ts_graph.variable_name_to_nodes), {'A': [ts_graph['A']], 'B': [ts_graph['B']], 'C': [ts_graph['C']]}
+            dict(ts_graph._variable_name_to_nodes), {'A': [ts_graph['A']], 'B': [ts_graph['B']], 'C': [ts_graph['C']]}
         )
+
+    def test_remove_non_existent_node_raises(self):
+        input_nodes = ['A lag(n=2)', 'A lag(n=1)', 'B']
+        ts_graph = TimeSeriesCausalGraph(input_list=input_nodes, fully_connected=False)
+
+        ts_graph._remove_node_from_cache(ts_graph.get_node(ts_graph._NodeCls.identifier_from('B')))
+        with self.assertRaises(ValueError):
+            ts_graph.delete_node('B')