+++ changelog.md
@@ -2,9 +2,8 @@
 
 ## NEXT
 
-- Added `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.add_node_to_cache` and
-  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.remove_node_from_cache` to avoid having to loop
-  through the nodes of the graph multiple times.
+- Improved efficiency of `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.get_nodes_at_lag` and
+  `cai_causal_graph.time_series_causal_graph.TimeSeriesCausalGraph.get_nodes_for_variable_name` by adding caching.
 
 ## 0.4.6